WinPCAP deps - bare minimum to build sniffer
This is probably not the right way to do distribute the libs and
headers we need, so i defer to the professional developers and
licensing folks on that. However, this is the "sufficient to build"
approach to providing everything needed to build sniffer on target.

Please do not merge this, it needs review and potentially more work
in the form of a library dealing with driver loading/unloading and
cleanup.